# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LIU Connect API",
    "description": "API specification with example payloads for testing in Postman, Swagger UI, or similar tools.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from sign-in or refresh. Use: Authorization: Bearer <accessToken>"
      }
    },
...
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "summary": "List posts",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "communityId", "in": "query", "schema": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440001" } },
          { "name": "authorId", "in": "query", "schema": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440000" } }
        ],
        "responses": {
          "200": { "description": "List of posts" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create post",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content", "media"],
                "properties": {
                  "content": { "type": "string", "example": "Hello world! This is my first post." },
                  "communityId": { "type": "string", "nullable": true, "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440001" },
                  "media": {
                    "type": "array",
                    "items": { "type": "string" },
                    "maxItems": 4,
                    "example": ["https://example.com/image1.jpg"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Post created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/comments": {
      "post": {
        "summary": "Create comment",
        "tags": ["Comments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["postId", "content", "media"],
                "properties": {
                  "postId": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440002" },
                  "content": { "type": "string", "example": "Great post!" },
                  "media": {
                    "type": "array",
                    "items": { "type": "string" },
                    "maxItems": 4,
                    "example": []
                  },
                  "parentCommentId": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440003", "description": "For replies" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Comment created" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Post not found" }
        }
      },
      "get": {
        "summary": "List comments",
        "tags": ["Comments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "postId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440002" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "userId", "in": "query", "schema": { "type": "string", "format": "uuid", "example": "550e8400-e29b-41d4-a716-446655440000" } },
          { "name": "parentCommentId", "in": "query", "schema": { "type": "string", "format": "uuid", "description": "Get replies to this comment" } }
        ],
        "responses": {
          "200": { "description": "List of comments" },
          "400": { "description": "postId required" },
          "404": { "description": "Post not found" }
        }
      }
    },
...
    "/api/communities": {
      "get": {
        "summary": "List communities",
        "tags": ["Communities"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string", "example": "tech" } }
        ],
        "responses": {
          "200": { "description": "List of communities" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create community",
        "tags": ["Communities"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 2, "example": "Tech Enthusiasts" },
                  "description": { "type": "string", "nullable": true, "example": "A place for tech lovers" },
                  "avatar_url": { "type": "string", "nullable": true, "example": "https://example.com/community.jpg" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Community created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "summary": "List notifications",
        "tags": ["Notifications"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of notifications" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}